Move to topology package
