+++ cmd/root.go
@@ -96,6 +96,10 @@ func NewRootCmd(cli Cli) *cobra.Command {
 
 func InitOptions(cli Cli, opts GlobalOptions) {
 	// color mode
+	if opts.Color == "" && os.Getenv("NO_COLOR") != "" {
+		// https://no-color.org/
+		opts.Color = "never"
+	}
 	switch strings.ToLower(opts.Color) {
 	case "auto", "":
 		if !cli.Out().IsTerminal() {
